+++ bot/store/nmm.js
@@ -3,17 +3,22 @@
  * Load new music monday in memory
  * and update it once an hour
  */
+const _private = require(process.cwd() + '/private/get'); 
+const settings = _private.settings;
+
 const schedule = require('node-schedule');
 const spreadsheet = require(process.cwd() + '/bot/utilities/spreadsheet');
 
+
 function loadIntoMemory(bot) {
-  spreadsheet.loadSheet('1R6zwrTjAyNQCTX2DAzgytL5xsAEji3Et7z5HrKxAg5g', 3)
+  spreadsheet.loadSheet(settings.spreadsheets.new_music_monday, 3)
     .then(function(rows){
-      return spreadsheet.getNext(rows, 'month');
+      // console.log(rows);
+      return spreadsheet.getNext(rows, 'date');
     })
     .then(function(row){
       bot.sheetsData = bot.sheetsData || {};
-      bot.sheetsData.nnm = row;
+      bot.sheetsData.nmm = row;
     })
     .catch(function(err){
       bot.log('error', 'BOT', err);